requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
cloudinary==1.38.0
pypdf==4.0.1